// src/GG.Portafolio.Site/Npm/TypeScript/Types/ElementRow.ts
import { TimeCalculated } from './TimeCalculated';

export type ElementRow = {
    row: HTMLDivElement;
    timeCalculated: TimeCalculated;
    firstColumn: HTMLDivElement;
    secondColumn: HTMLDivElement;
    thirdColumn: HTMLDivElement;
    fourthColumn: HTMLDivElement;
    fivethColumn: HTMLDivElement;
}